refactor(core): 拆分了 auth 相关内容 (#246)
