+++ noj-core/tests/lib/helper.ts
@@ -200,7 +200,9 @@ export async function createUserToken(
   );
   const { users, userRoles } = await import("../../src/db/schema.ts");
   const { signToken } = await import("../../src/lib/jwt.ts");
-  const { ensureRbacSeeds } = await import("../../src/services/seed/seed-rbac.ts");
+  const { ensureRbacSeeds } = await import(
+    "../../src/services/seed/seed-rbac.ts"
+  );
   // deno test 每个文件独立模块图：PGlite 模式下必须先引导 Schema，
   // 否则 users/roles 等表不存在（PG 模式无操作）。
   await ensurePGliteSchemaForTest();
