+++ noj-core/tests/services/problem-field-guard.test.ts
@@ -23,7 +23,10 @@ import {
 import { AppError, ForbiddenError } from "../../src/lib/errors.ts";
 import { ROOT_USER_ID } from "../../src/lib/constants.ts";
 import { ensureRbacSeeds } from "../../src/services/seed/seed-rbac.ts";
-import { createProblem, updateProblem } from "../../src/services/problems/problems.ts";
+import {
+  createProblem,
+  updateProblem,
+} from "../../src/services/problems/problems.ts";
 import { importProblemBundle } from "../../src/services/problems/problem-bundle.ts";
 import {
   resetSetting,
