+++ noj-core/tests/services/rbac.test.ts
@@ -59,7 +59,9 @@ for (
 }
 
 // 确保 RBAC 种子数据已加载（幂等）
-const { ensureRbacSeeds } = await import("../../src/services/seed/seed-rbac.ts");
+const { ensureRbacSeeds } = await import(
+  "../../src/services/seed/seed-rbac.ts"
+);
 await ensureRbacSeeds();
 
 // 获取预置角色 ID
